pyproject.toml
[project]
name = "acidni-support"
version = "1.0.0"
description = "Centralized support and feedback collection API for all Acidni apps"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Proprietary" }
authors = [
    { name = "Acidni LLC", email = "jamieson@acidni.net" }
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "azure-identity>=1.19.0",
    "azure-cosmos>=4.9.0",
    "azure-keyvault-secrets>=4.9.0",
    "httpx>=0.28.0",
    "pyyaml>=6.0.2",
    "bleach>=6.2.0",
    "python-multipart>=0.0.18",
    "applicationinsights>=0.11.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
